增加了草稿功能(给文章设置可见权限)

+++ src/main/java/com/chengfy/blog/controller/webentity/PostInfoEntity.java
@@ -7,9 +7,9 @@ import lombok.Data;
 @Data
 public class PostInfoEntity<T extends BaseModel> {
     private T info;     // Category, Tag, Section
-    private int postCount;
+    private Long postCount;
 
-    public PostInfoEntity(T info, int postCount) {
+    public PostInfoEntity(T info, Long postCount) {
         this.info = info;
         this.postCount = postCount;
     }
